Move helper functions to AppleCrashReport class, Lookup slide value in debug images
