+++ src/sentry/lang/native/utils.py
@@ -133,125 +133,3 @@ def get_sdk_from_apple_system_info(info):
         'version_minor': system_version[1],
         'version_patchlevel': system_version[2],
     }
-
-
-def get_apple_crash_report(threads, context, debug_images, symbolicated=False):
-    rv = []
-    rv.append(_get_meta_header(context))
-    rv.append(get_threads_apple_string(threads, symbolicated))
-    rv.append(get_binary_images_apple_string(
-        debug_images,
-        context,
-        symbolicated
-    ))
-    return '\n\n'.join(rv) + '\n\nEOF'
-
-
-def _get_meta_header(context):
-    return "OS Version: {} {} ({})\nReport Version: 104".format(
-        context.get('os').get('name'),
-        context.get('os').get('version'),
-        context.get('os').get('build'),
-    )
-
-
-def get_threads_apple_string(threads, symbolicated=False):
-    rv = []
-    for thread in threads:
-        thread_string = get_thread_apple_string(thread, symbolicated)
-        if thread_string is not None:
-            rv.append(thread_string)
-    return "\n\n".join(rv)
-
-
-def get_thread_apple_string(thread, symbolicated=False):
-    rv = []
-    stacktrace = thread.get('stacktrace')
-    if stacktrace is None:
-        return None
-    if stacktrace:
-        frames = stacktrace.get('frames')
-        if frames:
-            i = 0
-            for frame in reversed(frames):
-                frame_string = _convert_frame_to_apple_string(
-                    frame=frame,
-                    number=i,
-                    symbolicated=symbolicated
-                )
-                if frame_string is not None:
-                    rv.append(frame_string)
-                    i += 1
-
-    if len(rv) == 0:
-        return None  # No frames in thread, so we remove thread
-    thread_string = 'Thread {} name: {}\n'.format(thread['id'],
-        thread['name'] and thread['name'] or ''
-    )
-    if thread['crashed']:
-        thread_string += 'Thread {} Crashed:\n'.format(thread['id'])
-    return thread_string + "\n".join(rv)
-
-
-def _convert_frame_to_apple_string(frame, number=0, symbolicated=False):
-    if frame.get('instruction_addr') is None:
-        return None
-    slide_value = _get_slide_value()
-    instruction_addr = slide_value + int(frame['instruction_addr'], 16)
-    image_addr = slide_value + int(frame['image_addr'], 16)
-    offset = ''
-    if frame['image_addr'] is not None and not symbolicated:
-        offset = ' + {}'.format(
-            instruction_addr - slide_value - int(frame['symbol_addr'], 16)
-        )
-    symbol = hex(image_addr)
-    if symbolicated:
-        file = ''
-        if frame.get('filename') and frame.get('lineno'):
-            file = " ({}:{})".format(
-                frame['filename'],
-                frame['lineno']
-            )
-        symbol = "{}{}".format(
-            frame['function'],
-            file
-        )
-    return "{}{}{}{}{}".format(
-        str(number).ljust(4, " "),
-        frame['package'].rsplit('/', 1)[-1].ljust(32, " "),
-        hex(instruction_addr).ljust(20, " "),
-        symbol,
-        offset
-    )
-
-
-def _get_slide_value():
-    # TODO(hazat): this value is for 64bit devices (slide value)
-    # 0x0000000100000000 need to find out how to get it
-    # strangly it also works without slide_value
-    return 0
-    # return int('0x0000000100000000', 16)
-
-
-def get_binary_images_apple_string(debug_images, contexts, symbolicated=False):
-    # We dont need binary images on symbolicated crashreport
-    if symbolicated or debug_images is None:
-        return ''
-    binary_images = map(lambda i:
-        _convert_debug_meta_to_binary_image_row(i, contexts),
-        sorted(debug_images, key=lambda i: int(i['image_addr'], 16)
-    ))
-    return "Binary Images:\n" + "\n".join(binary_images)
-
-
-def _convert_debug_meta_to_binary_image_row(debug_image, contexts):
-    slide_value = _get_slide_value()
-    image_addr = int(debug_image['image_addr'], 16) + slide_value
-    return "{} - {} {} {}  <{}> {}".format(
-        hex(image_addr),
-        hex(image_addr + debug_image['image_size'] - 1),
-        debug_image['name'].rsplit('/', 1)[-1],
-        contexts['device']['arch'],
-        debug_image['uuid'].replace('-', '').lower(),
-        debug_image['name']
-    )
